Refactor: Node.js
changed from integer to double in favor for more precision

+++ src/utils/Node.java
@@ -1,27 +1,27 @@
 package utils;
 
 public class Node {
-	
-	private int x,y;
-	
-	public Node(int x, int y) {
+
+	private double x, y;
+
+	public Node(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 }
